requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
black==22.3.0
brotlipy==0.7.0
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
datasets==2.4.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
executing==1.0.0
fairscale==0.4.8
filelock==3.8.0
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.9.1
hydra-core==1.2.0
iopath==0.1.9
ipykernel==6.15.2
ipython==8.4.0
ipywidgets==8.0.1
jedi==0.18.1
joblib==1.1.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
multiprocess==0.70.13
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.2.3
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
portalocker==2.5.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==5.1
pyzmq==23.2.1
regex==2022.8.17
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seqeval==1.2.2
stack-data==0.5.0
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.6.7
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers @ git+https://github.com/huggingface/transformers.git@8c4a11493f2c7a9153d1711150a8830b96eacd2e
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
widgetsnbextension==4.0.2
xxhash==3.0.0
yacs==0.1.8
yarl==1.8.1
